pyproject.toml
[project]
name = "deploy-langchain"
version = "0.1.0"
description = ""
authors = [
    {name = "nahye",email = "knh8783@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "python-dotenv (>=1.1.1,<2.0.0)",
    "langchain (>=0.3.26,<0.4.0)",
    "langchain-openai (>=0.3.27,<0.4.0)",
    "langchain-community (>=0.3.26,<0.4.0)",
    "pypdf (>=5.7.0,<6.0.0)",
    "gradio (>=5.35.0,<6.0.0)",
    "gradio-pdf (>=0.0.22,<0.0.23)",
    "faiss-cpu (>=1.11.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "deploy_langchain", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
